Adding greater precision to setting representive
The changes are more precise in nature and leave less room for
oddness.

The problem this is addressing is on occassion, when you would select
the manuscript the value would not be set. My suspicion is that not
all of the update queries were firing (the Rails log confirmed this).
As to why they weren't? That is beyond me.

However this setup now leverages a common entry point for making
changes to the underlying attribute. Here's hoping it works.