Tweaking method signature
Instead of requiring requested_by on submit, I believe it is stronger
to require requested_by on initialization of the form. This means our
submit method is parameter free; And thus we are not changing
contextual state of the form after initialization.

What follows, with all of these commits, is the changes to make that
happen.