Rebuilding binstubs
I was encountering the following error.

```console
Could not find rake-10.4.2 in any of the sources
Run `bundle install` to install missing gems.
error: failed to push some refs to
  'https://github.com/ndlib/sipity.git'
```

And based on searching, I encountered the following thread.

http://stackoverflow.com/questions/27879119/rails-s-could-not-find-rake-10-4-2-in-any-of-the-sources-bundlergemnotfound

The solution that I attempted was:

```console
$ bundle --binstubs
```

It appeared to solve the issue. Note, these are only changes based on
existing bin files. There are oodles more bin files that were
generated.