Reducing chance of Airbrake swallowing exceptions
For a POST request that included an Attachment, processed by dragonfly,
an exception was encountered in our application logic. That exception
was in turn caught and handled via Airbrake. As part of Airbrake's
cleanup, it attempted to send #collect to the attachment. The
attachment, however, was already closed so #collect failed and threw
an exception.

This thrown exception, IOError, in turn obliterated the original
exception that was thrown by our application. It should be noted
that the current work around is a fork of the v4.3.0..master work
of Airbrake. Once issue 396 is resolved and released in a gem, we
can go back to using the default Airbrake gem.

Related to markevans/dragonfly#271
Related to markevans/dragonfly#271
Related to airbrake/airbrake#397
Related to airbrake/airbrake#396