Adding explicit require for Array.wrap
At present the .rsepc requires the Sipity's spec/rails_helper. In doing
so, all tests now load the entire Rails ecosystem. This results in very
slow tests for many of the unit tests. The problem remains hidden if
something like Spring is running; But Spring is a nasty kludge of a fix
that introduces all kinds of problems.