Moving away from Warden with authentication layer
Warden provides a quick short-circuit for jumping out of the
application when we don't have an authenticated user. However tying
into Warden is an opaque and poorly described process.

So I'm stepping away from that and towards something else. There is
an implied refactor that needs to happen:

See #887 and #888 for reference