Addressing a Rails pain point
At some point, the Sipity::Models were not being loaded; So the
use_relative_model_naming? directive was not being called. The problem
showed up when submitting the `state_advancing_action_confirmation`
HTML page. The form's `as:` was resolving as `sipity_models_work`
when in order for it to function it should be `work`.