Leveraging power converter to ensure common param
Prior to this commit, if I received roles, I was not verifying that
the roles were in fact roles. Then I was comparing directly the
role name vs. the given role. In unit tests this works reasonably
well. However in integrations I was passing the method a list of
roles then comparing the roles vs. the returned name.

This proved futile as I was not comparing apples to apples. So the
following change ensures that I'm comparing the proper thing.