Differentiating Cogitate development and test env
NOTE: cogitate_services_tokenizer_public_password is different for
development and test environment. It is assumed that the development
environment is connecting to a "live server" whereas test is often stubbing out
the connection. The most important thing is that the
`cogitate_services_tokenizer_public_password` is correct. The
`cogitate_services_tokenizer_public_password` is the public key for the remote
service. The cogitate token is signed with a corresponding private key and to
detokenize via JWT, the public key must match.