Adding sequestering state for ETD workflow
Because at this point in time, we have marked objects as ready for
ingest; However they are not ready as they need to either go through
cataloging or be marked as ready for immediate ingest with delayed
cataloging.

The plan is to change the data migration, which has yet to run, to
move items from Ready for Ingest to the temporary, though more
appropriate 'grad_school_approved_but_waiting_for_routing'.