Adding a bogus webhook callback for bulk ingest
In attempting to troubleshoot issues with character encoding of the
bulk ingest process, I was reviewing the data output from the various
jobs that were ran.

It was impossible to determine from which environment the batch request
originated. As a stop gap solution, I'm adding the supported WEBHOOK
behavior for the process. This will provide configuration information
for the job as it runs.