Updating ULRA's workflow completion process
Instead of having two different todo items within the workflow, I'm
opting to have students choose to sign off and faculty. This sends the
work into an interstitial state that the other person can work on.

To see the effect of this, compare the before and after processing
diagrams.

To get the processing diagram for this commit use the following:

```console
$ rake bootstrap
$ rails r scripts/commands/generate_state_machine_diagrams.rb
$ open artifacts/state_machines/ulra_submission_processing.dot -a Graphviz
```

Related to #949