Extracting a common email class builder method
Prior to this commit the EtdMailer had a heavy load of duplicate logic.
It wasn't as declarative as it could be.

This commit extracts a builder method that allows for the predictable
generation of the EtdMailer and an upcoming UlraMailer object. The
predictability is desirable from an validation and proper composition
stand point.

It should be noted that a few oddities are introduced with the builder,
namely the view paths in partials gets rather weird (i.e. see comments
in the builder for more details).

By extracting a function responsible for building the emailer in a
consistent manner, I can swap out the details of the actual email
subsystem. This could've been done by creating a subclass of the
ActionMailer::Base and adding the configuration options there. But then
we have an explicit 3 levels of inheritance, which becomes nasty.

It should also be noted that by moving towards this composition, it would
be possible to express the email notifications in a configuration file.
One thing I've been thinking about is whether the entire workflow should
be articulated in some kind of localized data structure instead of the
database. But that is another thought for later.