Adding open and close dates for submission window
Because the submission windows are not something that remain open
forever. So we need a means of controlling that. This gets more
challenging when we need to account for submissions that are being
worked on within a "closed window".

I think the easiest assumption is that you cannot start a submission
in a window that is closed. However, you can continue to work on the
submission. If that becomes a problem I will need to discuss with each
work area what policy they want.