Adding submission window validation to starting
Given that submission windows are can be closed, I want to ensure that
we are not allowing a person to start a submission if the window is not
open.

The validation is being enforced, but I now need to ensure that I'm
finding all submission windows which are presently open and presenting
that as a button on the dashboard for a given work area.