Removing self_deposit work area
Instead of carrying these half-baked ideas along in the application
development cycle, I'm removing them in one commit. This way we can
restore that information if it becomes necessary or revisit at a later
date.

This is a decision to keep the testing environment smoother. To help
ease the refactoring of ULRA and ETDs while not having to consider a
third abstract work area concept.