Adding cataloger's ability to see after action
As I was clicking through the cataloging workflow, I encountered an
unauthorized error after I had submitted a state changing action. This
was because cataloging did not have the ability to see the record in
its new state.

This may be a stop gap solution as I should consider a response handler
for when you send an object to a new state and no longer need to see
it. A flash message and a thank you?