Adding Ulra::AttachProcessingHook
ULRA has requested that when the student marks their thesis file(s) as
complete, they want to receive an email. They do not, however, want to
receive that email before the submission has been submitted for review.

```gherkin
Given a user updates their attached thesiss files
and has indicated their thesis files are complete
When this processing hook fires
Then notifications should be delivered for the given entity and its
  current state
```

This is triggered from an action but delivers emails for the given
state because not all states will have emails that should be sent.