Adding the ULRA submission window and schema
I want to move towards submission windows described as schema. This
will greatly ease some of the database related tests that go along
with permissions.

Also note, that I'm adding a few states to the ULRA configuration to
reflect that I should be managing state changes with open and close
functionality.