Renaming Roles to verb to reflect action
A role is a responsibility to do things. That is to take actions. It is
easy to conflate a person's role with the groups to which they belong.
A group is a set of people. The association of group with role
indicates that the set of people have the same responsibilities.

By separating Group and Role, we expose a more rich system in which we
can talk about group membership separate from the group's
responsibility.

Another way to think of it is that a Group is a marco that expands to
represent people. A Role is a macro that expands to represent
responsibilities. In keeping them separate we can model more rich
relationships.