Updating Collaborator roles to be limited
The ETD Collaborator model is a bit different than the ULRA
Collaborator model. There is a more challenging "Person" concept that
needs addressing, but for now this will need to do.

The change now ensures that each WorkType is responsible for specifying
the possible roles allowed. Its a loose enforcement that could be
tightened up a bit with some form modeling for the collaborator.