Updating emails to explain cataloging completion
We want to communicate to the students that just because the grad
school has approved the work, there is still more to do:

* cataloging
* ingest

This means first, apologizing for a delay in those records that were
submitted but not sent through the rigors. Second explaining that will
receive an email when the cataloging is complete.