Adding a spec to verify work type configs
There are some assumptions here, but in verifying the configuration it
provides a bit of documentation about how to take the configuration and
translate that into the corresponding ruby objects.

This required publicizing and converting the parameters for a
previously private class method. I'm reusing that method to highlight
how the application looks up what object to use.