Working on WebhookWriter interface and test coverage
I spent a lot of time trying to figuring out how to test writing to
files. The method I ultimately decided upon separates the creation of
the output buffer from the action that writes the data. This lets you
use a StringIO object as a stand-in for a file in your test cases.