Adding show etds to ETD_INTEGRATORS
As per the requirements of OnBase integration, adding the ability for
the ETD_INTEGRATORS to see the various works that are in the system.
They don't need to see "new" submissions nor "ingested" submissions.

The general idea is that we will have a handful of ETDs that are
working their way through submission; Hopefully no more than 150. As
such, I'm not yet introducing pagination.

However, it may need to be introduced if the system gets bogged down
creating the complicated JSON object.