Updating Sipity callback handling of request body
Since deploying the Sipity v2016.16 (or the day after), the webhook
callbacks have been firing from the batch ingester, but were not being
processed via the callback mechanism.

After some debugging, the WEBHOOK posting was coming as part of the
request's raw body (as per the documentation).

I ran the test I wrote against the v2016.16 codebase, and the test
failed. So something odd has happened.

Regardless, this code change now accounts for data posted as both
query parameters and request raw body.

Also adding a hound/rubocop exclusion for method complexity; This is a
compact method that is doing a few things, but I believe it remains
legible.