Removing JS error
Prior to this commit, the following was true:

```javascript
$checkbox[0] == undefined
```

This meant that `$checkbox[0].checked` was throwing an error. By
changing the code to check if its defined, I'm ensuring we don't
encounter an error in behavior.

I believe this to be the result of @b5f56b0 and not @51f2254.